Ensure two-step aggregation approximates one-step aggregation well
The two-step aggregation approach we employ to reduce the computational
cost of calculating buffer features for each individual building, and
instead calculate buffer features for the hexagonal grid cell in which
the building is located, presents some challenges:

It does not allow for exact calculation of some aggregates (e.g. std of
footprint area). We need approximate these by calculating the mean of
the standard deviations of each grid cell. Others can be calculated
exactly using another operation, for example the exact count of
buildings in a large buffer can also be calculated by counting buildings
in smaller buffers and suming these counts afterwards (i.e. the count
operation needs to be mapped to a sum operation in our two-step
aggregation approach).

We are explicit about the operations supported by our two-step
aggregation to avoid inadvertently incorrect aggregation.