One-hot encode closest OSM building type
Also, determine closest OSM building type and building height
separately, as the closest building may not have both attributes.
Instead, we independently determine the nearest building that has a
height attribute and the nearest building that has a type attribute -
even if they are two different buildings and other buildings without
these attributes are located even closer to the building of interest.